requirements.txt
# AnsFlow CI/CD Platform Requirements
# Python 依赖总览文件 - 包含所有服务的核心依赖

# ===========================================
# Django Service Dependencies
# ===========================================
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0.0
django-extensions>=3.2.0
djangorestframework-simplejwt>=5.2.0
django-environ>=0.10.0
django-redis>=5.2.0
celery>=5.3.0
redis>=4.5.0

# API Documentation
drf-spectacular>=0.26.0

# Encrypted fields
django-encrypted-model-fields>=0.6.5

# Database
mysqlclient>=2.1.0
PyMySQL>=1.0.0

# ===========================================
# FastAPI Service Dependencies  
# ===========================================
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
websockets>=11.0.0
aioredis>=2.0.0
aio-pika>=9.0.0

# ===========================================
# Shared Dependencies
# ===========================================
# HTTP Clients
httpx>=0.24.0
requests>=2.31.0
aiohttp>=3.8.0

# Database ORM
SQLAlchemy>=2.0.0
alembic>=1.11.0

# Message Queue
pika>=1.3.0
kombu>=5.3.0

# Monitoring & Logging
prometheus-client>=0.17.0
structlog>=23.0.0
sentry-sdk>=1.28.0

# ===========================================
# Development Dependencies
# ===========================================
# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=19.0.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.4.0
pre-commit>=3.3.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.1.0
mkdocs-mermaid2-plugin>=1.1.0

# ===========================================
# Security
# ===========================================
cryptography>=41.0.0
PyJWT>=2.8.0
passlib>=1.7.4
bcrypt>=4.0.0

# ===========================================
# Utilities
# ===========================================
python-dateutil>=2.8.0
pytz>=2023.3
click>=8.1.0
rich>=13.0.0
tenacity>=8.2.0
jinja2>=3.1.0
PyYAML>=6.0.0

# ===========================================
# Image Processing (Optional)
# ===========================================
Pillow>=10.0.0

# ===========================================
# File Handling
# ===========================================
python-magic>=0.4.27
python-slugify>=8.0.0

# ===========================================
# Email
# ===========================================
django-anymail>=10.0

# ===========================================
# Deployment & Infrastructure
# ===========================================
gunicorn>=21.0.0
whitenoise>=6.5.0
dj-database-url>=2.0.0

# Docker
docker>=6.1.0
docker-compose>=1.29.0

# Kubernetes
kubernetes>=27.0.0

# AWS (Optional)
boto3>=1.28.0
botocore>=1.31.0

# Configuration and Data
PyYAML>=6.0
python-dotenv>=1.0.0
configparser>=5.3.0
